Move regression tests to root
